NBA_Data_Core/python/requirements.txt
pandas
python-dotenv
pyodbc
prophet
statsmodels
scikit-learn